ISSUE-3 scaffolding
