+++ src/main/scala/com/utn/tacs/Main.scala
@@ -4,5 +4,5 @@ import cats.effect.{ExitCode, IO, IOApp}
 
 object Main extends IOApp {
   def run(args: List[String]) =
-    Tacs1c2021Server.stream[IO].compile.drain.as(ExitCode.Success)
+    Server.stream[IO].compile.drain.as(ExitCode.Success)
 }
