+++ src/main/scala/com/utn/tacs/Server.scala
@@ -1,6 +1,8 @@
 package com.utn.tacs
 
 import cats.effect.{ConcurrentEffect, Timer}
+import com.utn.tacs.domain.auth.Auth
+import com.utn.tacs.infrastructure.endpoint.{AuthEndpoints, DeckEndpoints}
 import org.http4s.server.Router
 //import cats.implicits._
 import fs2.Stream
@@ -10,14 +12,14 @@ import org.http4s.server.blaze.BlazeServerBuilder
 import org.http4s.server.middleware.Logger
 import scala.concurrent.ExecutionContext.global
 
-object Tacs1c2021Server {
+object Server {
 
   def stream[F[_] : ConcurrentEffect](implicit T: Timer[F]): Stream[F, Nothing] = {
     val authAlg = Auth.impl[F]
 
     val httpApp = Router(
-      "/auth" -> Tacs1c2021Routes.authRoutes[F](authAlg),
-      "/decks" -> Tacs1c2021Routes.decksRoutes()
+      "/auth" -> AuthEndpoints.authRoutes[F](authAlg),
+      "/decks" -> DeckEndpoints.decksRoutes(),
     ).orNotFound
 
     val finalHttpApp = Logger.httpApp(true, false)(httpApp)
@@ -25,7 +27,7 @@ object Tacs1c2021Server {
     for {
       //      client <- BlazeClientBuilder[F](global).stream
       exitCode <- BlazeServerBuilder[F](global)
-        .bindHttp(8080, "0.0.0.0")
+        .bindHttp(8080, "127.0.0.1")
         .withHttpApp(finalHttpApp)
         .serve
     } yield exitCode
